+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2020-06-16
+### Changes
+- Adds ability to set more config values like cookie domain, cookies paths etc.
+- Adds ability to connect to https instance
+
 ## [1.0.1] - 2020-06-14
 ### Fixed
 - Converts frontendSDK info into a map[string]interface{} when returning it